fixup! syslog: add basic syslog implementation
